UI margin fix and config pixels are now in double instead of ints

+++ QuickViewFile/Models/ConfigModel.cs
@@ -4,9 +4,9 @@
     {
         public int MaxSizePreviewKB { get; set; } = 50;
         public string ImageStretch { get; set; } = "Uniform";
-        public int PreviewHeight { get; set; } = 800;
-        public int PreviewWidth { get; set; } = 1200;
-        public int ZoomStep { get; set; } = 50;
+        public double PreviewHeight { get; set; } = 800;
+        public double PreviewWidth { get; set; } = 1200;
+        public double ZoomStep { get; set; } = 50;
         public string TextPreviewWordWrap { get; set; } = "NoWrap";
     }
 }
